requirements.txt
uvicorn
pydantic
pymongo
requests
fastapi
aiokafka
loguru
pytest
pytest-cov
pytest-asyncio
motor
pydantic[email]
python-dotenv